pyproject.toml
[project]
name = "ai-agent-workshop"
version = "0.1.0"
description = "AI Agent Workshop for Beginners"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain",
    "langchain-openai",
    "langgraph",
    "litellm",
    "openai",
    "python-dotenv",
    "streamlit",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
